+++ client/package.json
@@ -9,6 +9,9 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@emotion/react": "^11.10.6",
+    "@emotion/styled": "^11.10.6",
+    "@mui/material": "^5.11.13",
     "@picocss/pico": "^1.5.7",
     "axios": "^1.3.1",
     "bootstrap": "^5.2.3",
