+++ client/src/pages/index.js
@@ -3,4 +3,3 @@ export { default as Contact } from "./contact/Contact";
 export { default as Login } from "./login/Login";
 export { default as Products } from "./products/Products";
 export { default as Detail } from "./detail/Detail";
-export { default as LandingPage } from "./landingPage/LandingPage";
